feat: UserWorking 모델 추가 및 생성한 모델 admin 등록 #7
