+++ project/alert/models.py
@@ -1,12 +1,13 @@
-from datetime import timezone
+from django.utils import timezone
 
 from django.db import models
 
-from project.user.models import User
-
+from django.contrib.auth import get_user_model
+User = get_user_model()
 
 class Alert(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     message = models.CharField(max_length=300)
     openChatURL = models.URLField(null=True)
     regTime = models.DateTimeField(timezone.now())
+    is_read = models.BooleanField(default=False)
