+++ project/artist/models.py
@@ -1,11 +1,8 @@
 from django.db import models
 
-from project.user.models import User
-
-
 class Artist(models.Model):
     artistName = models.CharField(max_length=50)
-    members = models.CharField(max_length=200,null=True)
+    members = models.CharField(max_length=200, null=True)
     artistImage = models.ImageField()
 
     instagram = models.URLField()
@@ -16,4 +13,3 @@ class Artist(models.Model):
     color = models.CharField(max_length=30)
     fanName = models.CharField(max_length=30)
     agency = models.CharField(max_length=30)
-    fans = models.ManyToManyField(User, null=True)
