+++ project/post/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from .models import Post, Comment
 
-# Register your models here.
+admin.site.register(Post)
+admin.site.register(Comment)
