+++ project/user/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from .models import User, Artist
 
-# Register your models here.
+from .models import User
+
 admin.site.register(User)
-admin.site.register(Artist)
+
